Cleanup: unify boost validation rules
